requirements.txt
# flask and dash
Flask
dash
dash-bootstrap-components
Jinja2
flask-bcrypt 
python-dotenv
flask-cors



# data manipulation
pandas
openpyxl
xlrd
XlsxWriter
numpy

# caching and performance
Flask-Caching



# Database (MySQL)
Flask-SQLAlchemy
pymysql

# deployment
gunicorn


# setup
setuptools